Updated profile_mysqli.php to keep it more in sync with the profile.php file.
Use databaseinfo.php file. Escape parcelname in classified_update. Open
(and close) database in one place. Use global $link to open database.

One user really wanted a version using mysqli so this file was not deleted.
This file is obsolete with the update of profile.php to use PDO so it may
not be kept up-to-date or be as fully tested as the main profile.php file.